+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Fixed
+- Fixed "Attempt to read property id on null" warning in get_view_url() when viewing
+  a single section page #109
+
 ## [5.0.1] - 2025-10-07
 ### Changed
 - Added tests to ensure compatibility with Moodle 5.1